# swagger.yaml
openapi: 3.0.0
info:
  title: User Management API For Alfred
  version: 1.0.0
  description: A simple CRUD API documented with Swagger

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /:
    get:
      summary: Welcome route
      description: Returns a welcome message
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: Welcome to the User Management API!

  /users:
    get:
      summary: Get all users
      description: Returns a list of all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      summary: Create a new user
      description: Adds a new user to the list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /users/{id}:
    get:
      summary: Get a user by ID
      description: Returns a single user by ID
      parameters:
        - in: path
          name: id
          required: true
          description: The user ID
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    put:
      summary: Update a user
      description: Updates a user's name by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    delete:
      summary: Delete a user
      description: Removes a user from the list by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alfred

    UserInput:
      type: object
      properties:
        name:
          type: string
          example: Chris